booking-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose
aiokafka
pydantic-settings